requirements.txt
caldav==1.4.0
google_api_python_client==2.169.0
Kivy==2.3.1
protobuf==6.31.0
python-dotenv==1.1.0
